Refactor image.tsx
