+++ frontend/src/components/NavBar/PrintImage/image.tsx
@@ -18,7 +18,6 @@ import {
 } from '@material-ui/core';
 import GetAppIcon from '@material-ui/icons/GetApp';
 import mask from '@turf/mask';
-import { legendListId } from 'components/MapView/Legends';
 import html2canvas from 'html2canvas';
 import { debounce } from 'lodash';
 import { jsPDF } from 'jspdf';
@@ -47,6 +46,9 @@ import { getBoundaryLayerSingleton } from 'config/utils';
 import { LayerData } from 'context/layers/layer-data';
 import { AAMarkersSelector } from 'context/anticipatoryActionStateSlice';
 import { useAAMarkerScalePercent } from 'utils/map-utils';
+import LegendItemsList, {
+  legendListId,
+} from 'components/MapView/Legends/LegendItemsList';
 import {
   dateRangeSelector,
   layerDataSelector,
@@ -159,6 +161,7 @@ function DownloadImage({ classes, open, handleClose }: DownloadImageProps) {
   const printRef = useRef<HTMLDivElement>(null);
   const overlayContainerRef = useRef<HTMLDivElement>(null);
   const titleOverlayRef = useRef<HTMLDivElement>(null);
+  const footerOverlayRef = useRef<HTMLDivElement>(null);
   const boundaryLayerState = useSelector(
     layerDataSelector(boundaryLayer.id),
   ) as LayerData<BoundaryLayerProps> | undefined;
@@ -247,25 +250,7 @@ function DownloadImage({ classes, open, handleClose }: DownloadImageProps) {
     setAdminBoundaryPolygon(masked as any);
   }, [data, selectedBoundaries, selectedBoundaries.length]);
 
-  const createFooterElement = (
-    inputFooterText: string = t(DEFAULT_FOOTER_TEXT),
-    width: number,
-    ratio: number,
-    fontSize: number,
-  ): HTMLDivElement => {
-    const footer = document.createElement('div');
-    // eslint-disable-next-line fp/no-mutation
-    footer.innerHTML = `
-      <div style='width:${
-        (width - 16) / ratio
-      }px;margin:8px;font-size:12px;padding-bottom:8px;font-size:${fontSize}px'>
-        ${inputFooterText}
-      </div>
-    `;
-    return footer;
-  };
-
-  const refreshImage = async (currentFooterText = footerText) => {
+  const refreshImage = async () => {
     /* eslint-disable fp/no-mutation */
     setElementsLoading(true);
     if (open && mapRef.current) {
@@ -290,7 +275,6 @@ function DownloadImage({ classes, open, handleClose }: DownloadImageProps) {
 
       const canvas = document.createElement('canvas');
       if (canvas) {
-        let footerTextHeight = 0;
         let scalerBarLength = 0;
         const scaleBarGap = 10;
 
@@ -315,76 +299,57 @@ function DownloadImage({ classes, open, handleClose }: DownloadImageProps) {
 
         // toggle legend
         const div = document.getElementById(legendListId);
-        if (div?.firstChild && legendScale > 0) {
-          const childElements = Array.from(div.childNodes).filter(
-            node => node.nodeType === 1,
-          ) as HTMLElement[];
-
-          const target = document.createElement('div');
-          target.style.width = '196px'; // 180px + 2*8px padding
-
-          childElements.forEach((li: HTMLElement, i) => {
-            const isLast = childElements.length - 1 === i;
-
-            const children = Array.from(li.childNodes).filter(
-              // node type 1 represents an HTMLElement
-              node => node.nodeType === 1,
-            ) as HTMLElement[];
-            const divContainer = children[0] as HTMLElement;
-
-            const contents = Array.from(divContainer.childNodes).filter(
+        if (0) {
+          if (div?.firstChild && legendScale > 0) {
+            const childElements = Array.from(div.childNodes).filter(
               node => node.nodeType === 1,
             ) as HTMLElement[];
 
-            const container = document.createElement('div');
-            container.style.padding = '8px';
-            container.style.paddingBottom = isLast ? '8px' : '16px';
-            target.appendChild(container);
+            const target = document.createElement('div');
+            target.style.width = '196px'; // 180px + 2*8px padding
 
-            const keepDivider = isLast ? 1 : 0;
+            childElements.forEach((li: HTMLElement, i) => {
+              const isLast = childElements.length - 1 === i;
 
-            contents
-              .slice(
-                0,
-                toggles.fullLayerDescription
-                  ? 6 - keepDivider
-                  : 4 - keepDivider,
-              )
-              .forEach(x => container.appendChild(x.cloneNode(true)));
-          });
+              const children = Array.from(li.childNodes).filter(
+                // node type 1 represents an HTMLElement
+                node => node.nodeType === 1,
+              ) as HTMLElement[];
+              const divContainer = children[0] as HTMLElement;
 
-          document.body.appendChild(target);
+              const contents = Array.from(divContainer.childNodes).filter(
+                node => node.nodeType === 1,
+              ) as HTMLElement[];
 
-          const c = await html2canvas(target, { useCORS: true });
-          context.drawImage(
-            c,
-            24,
-            24 + (titleOverlayRef.current?.offsetHeight || 0),
-            (target.offsetWidth * legendScale * ratio) / 100.0,
-            (target.offsetHeight * legendScale * ratio) / 100.0,
-          );
-          document.body.removeChild(target);
-        }
+              const container = document.createElement('div');
+              container.style.padding = '8px';
+              container.style.paddingBottom = isLast ? '8px' : '16px';
+              target.appendChild(container);
 
-        // toggle footer
-        if (footerTextSize > 0) {
-          const footer = createFooterElement(
-            currentFooterText,
-            activeLayers.width,
-            ratio,
-            footerTextSize,
-          );
-          document.body.appendChild(footer);
-          const c = await html2canvas(footer);
-          footerTextHeight = footer.offsetHeight;
-          context.drawImage(
-            c,
-            0 * ratio,
-            activeLayers.height - footer.offsetHeight * ratio,
-            footer.offsetWidth * ratio,
-            footer.offsetHeight * ratio,
-          );
-          document.body.removeChild(footer);
+              const keepDivider = isLast ? 1 : 0;
+
+              contents
+                .slice(
+                  0,
+                  toggles.fullLayerDescription
+                    ? 6 - keepDivider
+                    : 4 - keepDivider,
+                )
+                .forEach(x => container.appendChild(x.cloneNode(true)));
+            });
+
+            document.body.appendChild(target);
+
+            const c = await html2canvas(target, { useCORS: true });
+            context.drawImage(
+              c,
+              24,
+              24 + (titleOverlayRef.current?.offsetHeight || 0),
+              (target.offsetWidth * legendScale * ratio) / 100.0,
+              (target.offsetHeight * legendScale * ratio) / 100.0,
+            );
+            document.body.removeChild(target);
+          }
         }
 
         if (toggles.scaleBar) {
@@ -407,7 +372,7 @@ function DownloadImage({ classes, open, handleClose }: DownloadImageProps) {
               c,
               activeLayers.width - (scaleBarGap + elem.offsetWidth) * ratio,
               activeLayers.height -
-                (30 + (footerTextSize > 0 ? footerTextHeight : 0)) * ratio,
+                (30 + (footerOverlayRef.current?.offsetHeight || 0)) * ratio,
               html.offsetWidth * ratio,
               html.offsetHeight * ratio,
             );
@@ -426,7 +391,7 @@ function DownloadImage({ classes, open, handleClose }: DownloadImageProps) {
               activeLayers.width -
                 (scaleBarGap + imageWidth / 4 + scalerBarLength / 2) * ratio,
               activeLayers.height -
-                (110 + (footerTextSize > 0 ? footerTextHeight : 0)) * ratio,
+                (110 + (footerOverlayRef.current?.offsetHeight || 0)) * ratio,
               imageWidth,
               imageHeight,
             );
@@ -559,6 +524,29 @@ function DownloadImage({ classes, open, handleClose }: DownloadImageProps) {
                   {titleText && (
                     <div ref={titleOverlayRef} className={classes.titleOverlay}>
                       {titleText}
+                    </div>
+                  )}
+                  {footerTextSize > 0 && (
+                    <div
+                      ref={footerOverlayRef}
+                      className={classes.footerOverlay}
+                      style={{
+                        fontSize: `${footerTextSize}px`,
+                      }}
+                    >
+                      <div style={{ padding: '8px' }}>{footerText}</div>
+                    </div>
+                  )}
+                  {legendScale > 0 && (
+                    <div
+                      style={{
+                        position: 'absolute',
+                        top: 0,
+                        right: 0,
+                        zIndex: 2,
+                      }}
+                    >
+                      <LegendItemsList showOptions={false} />
                     </div>
                   )}
                   <div className={classes.mapContainer}>
@@ -827,6 +815,15 @@ const styles = (theme: Theme) =>
       textAlign: 'center',
       fontSize: '1.5rem',
       padding: '8px 0 8px 0',
+    },
+    footerOverlay: {
+      position: 'absolute',
+      bottom: 0,
+      left: 0,
+      zIndex: 2,
+      color: 'black',
+      backgroundColor: 'white',
+      width: '100%',
     },
     formControl: {
       width: '100%',
